// src/model/artist.ts
export type Artist = {
  id: number;
  slug: string;
  acf: {
    order: number;
    artist_name: string;
    artist_description: string;
    artist_description_eng: string;
    main_artist_image: string;
    artist_work_image_1: string;
    artist_work_image_2: string;
    artist_work_image_3: string;
    artist_work_image_4: string;
    artist_work_image_5: string;
    artist_work_image_6: string;
    artist_work_image_7: string;
    artist_work_image_8: string;
    artist_work_image_9: string;
    artist_work_image_10: string;
  };
};